Create a component to handle language selection
It feels easier if this would be in a section in the menu with the
other options both in mobile as in non-mobile (only the latter has
them in inside of the page's content now).  The component which will
help with language selection will help split off logic either way.