Transition to 'products.beef' when visiting 'products.index'
The default display for products should show 'products.beef'.  This is
handled by creating an index route which does the transition with the
'activate' hook.

There seems to be a bug when transitioning to 'products' when already
visiting the 'products' route.  We've yet to uncover why this isn't
operating as we'd expect.