Always fetch complete locations
The amount of locations received might not be complete.  In order to
ensure we have the full list, we now fetch them all on every call.

There are some optimizations which can be made in order to ensure we
have the full contents without fetching the data upon each request.
We'll leave that for later.