Approach for dynamically rendering delivery kinds
This approach allows us to dynamically render delivery kinds.  The
delivery kinds are discovered by URI and rendered with a slug.

Upon further inspection of the rendered contents, we expect these to
be static with a high chance of embedded contents.  However, we also
assume there's a fair chance that the individual route locations will
receive a page.  These are conflicting in terms of optimal
architecture.  We will work with the currently known elements, but
assume it may change over time, based on the ideas of the customer.