Render contents for location kind from backend
The backend now drives the content which should be shown for a
specific location kind.  This approach allows us to be fairly generic
about the contents, yet still drive the UI as we'd like.  It's not
perfect either, as dynamic elements can not easily be added in the
current construction.