New product card with visualization and such and abstractions
This is a messy commit.  Sorry, it should be cleaner.  I was trying to
get the product card in a state in which it could be demoed.  That's
the case right now, but the positioning is too far off stil.  The plus
and minus buttons aren't at the level they need to be either.

Note that for amount of items per package, we need plus/minus buttons
rather than a data-entry because a limited set of options is
possible.  We should figure something out for amount in grams too and
I think we can make something up for that as well.

Some animations have been added, all through CSS.  Nothing fancy going
on.  They can be improved.