Allow users to select the current offer
This meant extending the number component so it would cope with
disconnected objects.  The benefit of this approach is that the
logic is mostly contained to that fairly abstract component.

We've upgraded emberjs to 3.11.0 because we're running into some
issues which we don't have with Octane.  Mostly different behaviour of
Ember.  It is not our intention to figure out where Ember stands
exactly on this version, but rather to see how we can keep moving
forward towards Octane.