Implementing semi checkbox functionality for delevery method
The deleviry method uses checkboxes instead of radio buttons.  As only
one solution is allowed we're actually interested in using radio
buttons.  In order not to alter the style too much, we've stuck with
checkboxes and made them behave the way we need them to.

In our changes we've altered the CSS to drop a class which was always
set anyways.  We've also removed all the ASP related things.  The
dropdown methods will be changed as we can most likely turn them into
a map which allows them to select the desired location.  It may be
that we need to cater for something else than the map for
screenreaders too, but we'll see how that lands after we've
implemented the nicer strategy.