Frontend implementation for using the image service
The image service can be used to fetch images with the right
resolution.  This is a way of embedding the service in the frontend.
We notice that we have to execute a bunch of workarounds in order to
make the service work this way.  As such we'll likely go back to the
drawing board and update this code so it's much less hackish and
easier to read.

There are a few solutions.  The most common pattern is likely to drop
this logic into the model with a sensible set of default sizes which
can be used throughout the application.  That has its downsides too,
but it would be easier to read.  We can further abstract the
calculation and offer it as a separate method for other users.