Support product details for products without offers
Products which we can see will probably always have offers.  I don't
know.  Anyhow, this certainly doesn't hold during testing.  So
although we don't have a nice design for a product which currently
isn't on offer, we also don't want the site to crash on this
condition.

The condition is now semi supported.  It's not impressive, but it's an
improvement.