Only show warnings after pressing the next button on order info
We used to show warnings only when pressing the next button, however
there was an issue that when landing on the screen again, we would have
remembered to show the warning itself.  That's not nice and this commit
resets the state of the button press when navigating away from the
route.