Smoother number buttons
The text itself can't be selected anymore.  Options which don't exist hide
away (no more minus button when no option can be selected).  Hiding is fading
which means it works immediately but it's a bit smoother on the eyes appearing
and disappearing.  Text is centered in the middle to make it feel a bit more
weighted.  We turn the text on the button in the same color as the background
when it's disabled as the message seems to be visually implied.