Price calculation through promises and resources
This is probably not the best approach but it is one which was used in the app
before and it's not readily obvious what the best approach is anyhow.  This
emits a promise as well as a resource for indicating certain calculations, such
is the product and price of an order-line.