Only show products which belong to business when they're linked
If a business-entity is linked to the currently selected location, then we will
now only show the products which are available on that business entity's
location.

This structure will allow us to provide specific products at specific locations.
Our goal is to use this in a more broadly applied system, yet the frontend can
now filter out based on whichever rule is chosen elsewhere.